Add width selector for button block (#25999)
* Add sidebar width selector for Button block

Initial commit adding a width option for the Button block with
options at 25, 50, 75, 100% widths. By default none are applied
and the button is sized by its content.

* Persist width changes

* Apply styling through class name rather than inline

* Clean up CSS

* Allow toggling a custom width off

A width selection can now be toggled off by clicking the already
selected option.

* Update InspectorControls to use sentence case

Co-authored-by: Miguel Fonseca <miguelcsf@gmail.com>

* Remove max width from buttons with custom width set

PR #23168 overhauls alignment controls on the Buttons container and
allows for full width. For a Button inside a very large container,
in order to be set correctly to 100% it must be able to exceed any
configured max-width.

This commit does not remove the max-width for any button that does
NOT have a custom width percentage selected. This is so that
default buttons continue to be sized as they normally would.

* Apply `has-custom-width` class on frontend

Fixes a bug where the inner button was not being set correctly to 100%
width on the frontend, due to the missing classname.`

Co-authored-by: Staci Cooper <staci@Stacis-MacBook-Pro.local>
Co-authored-by: Miguel Fonseca <miguelcsf@gmail.com>