Include button margins in width calculations (#26781)
* Include button margins in width calculations

This allows, for example, two 50% width buttons to fit next to
each other on a single row. Left and right margins are removed
for buttons with width at 100%.

* Include margins for buttons with 100% width

Including the margins even for buttons with 100% width allows
them to line up correctly with buttons of other widths.

For example, two 50% buttons and one 100% button will take up two
rows. For the second row with the 100% button to be flush with
the row above it, it must also include a margin.

Co-authored-by: Staci Cooper <staci@Stacis-MacBook-Pro.local>