Site Editor: Add theme taxonomy to templates and template parts (#27016)
Replace theme meta with theme taxonomy. Register a taxonomy called wp_theme and use it to store the theme of the template and template parts.

A special value is used to mark templates that were created from a theme template file. This value is: _wp_file_based.

Note: you may need to clean the FSE database before using this.