Block Support: Separate opt in for font style and weight options (#26844)
Updates the recently added font appearance block support to separate the style and weight aspects in terms of opting in via block.json. This required:

* Updating the detection of support opt in within typography tools
* Updating the UI control to handle the situation when only one option is opted into

This change also includes adding translation for the combined UI option labels and extracting concerns of how the selections are stored within style attributes from the UI control.