BlockSwitcher: Fix crash due to null reference (#28122)
In specific situations the mapSelect() in BlockSwitcher can run before
the mapSelect() in BlockToolbar. This means that if one of the selected
blocks has been deleted then there is a null reference error in
BlockSwitcher.

One such situation is: if a block is selected, the viewport shrinks from
desktop to mobile, and then the block is deleted.