Remove border radius for root element (#28320)
The support declaration was named "border" but should have been named "borderRadius", hence it didn't work.

These styles are attached to the :root selector so I'm not sure it makes much sense to have this property. There's also the issue that there's no UI control for users to tweak this property via the global styles sidebar.

An alternative that was considered to fix this is renaming it to borderRadius.