Site Editor: Fix navigate-to-link error (#29239)
* Add guard when navigate to link is not on a page

In the context of full site editing, the editor may load blocks outside of the context of a page. For example, if a user opens the navigation sidebar and loads a  template part, there will be no page entity and no page metadata.

Instead, there will be template part metadata because we are loading a template part entity. The changes here allow navigate-to-links to be loaded in other contexts outside of a page.