Release Workflow: Allow triggering manually (#28138)
Allow manual triggering of the workflow to create a new release, per [this discussion](https://github.com/WordPress/gutenberg/pull/27488#issuecomment-757946026).

The end goal is for the release manager to run the entire release process from GitHub, and to minimize manual interaction.
- :hand: Release manager starts release process manually.
- :robot: GitHub workflow bumps version numbers in plugin files, builds Gutenberg plugin, creates release draft, and prefills it with release notes based on changelog script output.
- :hand: Release manager edits release draft, and eventually publishes it. This creates the tag.
- :robot: (Other) GitHub workflow prepends release notes to `changelog.txt`, uploads Gutenberg plugin and updated `changelog.txt` to WP.org plugin repo.

Functionally, this is as close as possible to the previous workflow. The major difference is that `changelog.txt` is only updated _after_ tagging. This is a result of using the GitHub release draft UI to edit and revise release notes. It's not possible to include the result of those changes _before_ tagging.

Release docs are updated to reflect these changes. The previously used local release script is removed.