[RNMobile] Image Block: Improve Text Entry for Long Alt Text (#29670)
* Create new AltTextSettings component

With this commit, a new AltTextSettings component has been introduced to replace the current TextControl that exists for entering an image's Alt text. This new component makes use of BottomSheet.SubSheet, so that users are taken to a new subsheet when entering alt text.

* Update text colour in footer note in styles.css

The current text in the alt text subsheet's footer doesn't meet WordPress' accessibility standards for colour contrast. This commit updates the colour from " $gray-30" to "$gray-text-min" in order to meet those standards.

* Increase min-height of altTextEditor

This commit increases the min-height of "altTextEditor" so that the editor is a larger height on iOS devices by default.

* Update border colours for both light and dark mode

This commit updates the component's border colours to match similar separators/borders throughout the app. "light-gray-400" is used for light mode and "gray-70" for dark mode.

* Update text colour in dark mode

This commit fixes a bug, where text entered in iOS appear black on black (therefore, unreadable) in dark mode.

* Update footer note styling in styles.css

With this commit, I've tweaked the font size and padding around the "altTextSetting" compont's footer note to more closely match up with the initial design.

* Set max height for the editor

This commit fixes a bug I spotted while testing, where it's possible to continue hitting "enter" and increasing the size of the editor without limitless, leading to the footer note being dragged off screen.

* Increase bottom padding of footer note

This commit increase the padding at the bottom of the footer note, in accordance with the design.

* Update colour of placeholder text

This commit updates the colour of the default placeholder text in the "AltTextSettings" component's TextInput. The purpose of defining this is to fix a bug in iOS, where the placeholder text appeared as black on black in dark mode.

* Styling tweaks to footnote and placeholder text

Two small changes to the "Alt Text Settings" components are included in this commit. The first increases the padding of the footnote to match the design and the second aligns the placeholder text to the top, which is necessary for Android devices.

* Update font-size of footnote

Changes the font-size of the footnote from 13px to 12px, based on feedback.

* Set defined height for altTextEditor

This commit sets a defined height of 80px for the altTextEditor component, in response to feedback and to remove complexity around the box expanding when text is entered.

* Update blue colour of link for both light and dark mode

The colour of the link is updated for both light and dark mode with this commit, based on feedback received.

* Update CHANGELOG.md

Update CHANGELOG.md with changes from this commit.

* Convert "AltTextSettings" component into more generic "BottomSheetTextControl" component

This commit, converts the "AltTextSettings" component into more generic "BottomSheetTextControl" component by removing references to "alt" from main component and passing down props.

* Created README.md for BottomSheetTextControl component

This commit introduces a README.md for the new BottomSheetTextControl component.

* Formatting tweaks for README.md

* Formatting tweaks for README.md

Fixes some issues with spacing.

* retrigger checks

* Simplify BottomSheetTextControl

This commit seeks to simplify the BottomSheetTextControl component by splitting out two more, smaller components from it: 1. smallFooterMessageControl 2. smallFooterMessageControlLink

* retrigger checks

* Correct merge conflict

* Update "SmallFooterMessageLinkControl" to "FooterMessageLink"

This commit updates the name of "SmallFooterMessageLinkControl" to "FooterMessageLink" in order to be clearer and more readable.

* Update "SmallFooterMessageLinkControl" to "FooterMessageLink"

This commit updates the name of "SmallFooterMessageLinkControl" to "FooterMessageLink" in order to be clearer and more readable.

* Delete small-footer-message-control-link.native.js

* Remove small-footer-message-control

* Update Footer Message Control

* Add README to FooterMessageLink component

This commit moves FooterMessageLink to a separate folder for the purpose of adding a README

* Revert accidental commit

This reverts an accidental commit, where I'd made changes in "link-settings/index.native.js" as part of general testing. I also remove "textAlign="left"" from a FooterMessageControl component in this file, as this is now the default.