Global Styles: Make the config resilient to changes to `safecss_filter_attr()` (#30888)
* Test remove_insecure_properties() isn't too strict

The CSS values in the theme object can not be unsafe, but we also
shouldn't be so strict that values that would be accepted by a CSS
parser are rejected.

* Harden remove_insecure_properties() to implementation detail changes

The method was assuming a particular implementation of the
safecss_filter_attr() function. In the future it could do something like
remove whitespace around the ":", or convert lowercase colours (#abcabc)
to uppercase (#ABCABC), which would break saving of Global Styles in
Gutenberg. In general it doesn't feel like we should assume filtering
functions will return the input unchanged.

Switching the semantics of the check from "style was unchanged" to
"style was not stripped" fits better with the description of the
safecss_filter_attr() function.