FSE: Add template loader unit tests (#31498)
The block template resolution algorithm is quite complex, as it has to cover quite a lot of different possible cases resulting from various combinations of
- WP's "old" template hierarchy (including custom page templates, and child themes)
- block templates.

Things become especially complex when there are e.g. "oldschool" PHP templates with higher specificity than the available block templates; or when a child theme has a PHP template with equal specificity as the parent theme's corresponding block template.

For more discussion on this, see https://github.com/WordPress/gutenberg/issues/31399. Examples of previous iterations that sought to refine the algorithms' behavior include #29026, #30599, #31123, and #31336. This PR's goal is to eventually cover all of those cases.