[RNMobile] Add "Set as Featured Image" Button to Image Block (Android Only) (#31705)
* Add setFeaturedImage() methods to bridge

* Add getSetFeaturedButton() function which returns a "set/remove featured| button

The function returns a button (in the form of a BottomSheet.Cell component) based on whether "isFeaturedImage" is true or false. The function called within the image block's settings.

In its current form, the button is unresponsive and only includes a label to "Set as Featured Image" or "Remove as Featured Image", depending on whether the current image is already featured or not.

* Wrap "androidOnly" conditional around "Set/Remove Featured Image" button

For now, the button will only work on Android devices, so an "androidOnly" conditional is necessary to avoid bugs/complications on iOS.

* Define button's CSS selectors for styling

The getStylesFromColorScheme hook is used to define CSS selectors for the button, depending on whether the device uses dark or light mode. The hook's "featuredButtonStyle" variable is then called via BottomSheet's labelStyle attribute.

* Update styles.native.scss with styling for featuredButton and featuredButtonDark

* Define CSS selector for "Set as Featured Image" button

This commit uses the "getStylesFromColorScheme" hook to set up CSS selectors for the "Set as Featured Image" button, as its style will vary slightly from the base button.

* Update styles.native.scss with styling for setFeaturedButton

This includes CSS for both the light mode and dark mode versions of the button.

* Define CSS selector for "Remove as Featured Image" button

* Update styles.native.scss with styling for removeFeaturedButton

* Set up "onSetFeatured" function for when button is pressed

This function will fire when the "Set/Remove as Featured Image" button is pressed and its functionality will be fleshed out in future commits.

* Import setFeaturedImage and call it from onSetFeatured

setFeaturedImage is imported from '@wordpress/react-native-bridge' and called via the onSetFeatured, as it'll be utilised when the button is pressed.

* Pass mediaId of new featured image into onSetFeatured() and setFeaturedImage()

When a featured image is being removed, the mediaId will be zero, and when it's being set it'll be the "attributes.id" of the image in the block.

* Add closeSettingsBottomSheet() to dismiss BottomSheet when button's pressed

* Update CHANGELOG.md

* Update reference to 'core/edit-post' to avoid store name literal

* Move androidOnly conditional

The conditional is moved to the main call for this.getSetFeaturedButton(), to prevent unecessary code from being load.

* Create "setFeaturedButton" and "removeFeaturedButton" consts

These consts are then returned by the component, with the purpose being to make the return statement more readable.

* Merge "featuredButton" and "setFeaturedButton" CSS styles

For better readability, this commit merges the separates selectors into one main "setFeaturedButton" selector.

* Add check for whether image's attribute.id is "undefined"

This check will prevent getSetFeaturedButton() running in cases where an image's attribute.id is undefined, as attempting to set an image without a valid id as feature would lead to a crash.

For more background, see: https://github.com/WordPress/gutenberg/pull/31705#issuecomment-841317965

* Create "canImageBeFeatured" boolean

* Replace some logic in "isFeaturedImage" boolean with new "canImageBeFeatured" boolean

* Remove unecessary fragments

* Add comment to give context to 'canImageBeFeatured' check

* Add MEDIA_ID_NO_FEATURED_IMAGE_SET const

With this commit, a MEDIA_ID_NO_FEATURED_IMAGE_SET constant is added to make the reason we pass a zero when a featured image is removed clearer, and also make it easier to change in the future if needed.

* Move "MEDIA_ID_NO_FEATURED_IMAGE_SET" constant from within function to top of file

* Correct mixup between "setFeaturedButton" and "removeFeaturedButton" functions

These functions had accidentally been given the oppositive names (that is, the function that removes a featured image was named "setFeaturedButton" and vice versa). This commit corrects that mistake.

* Add 'MEDIA_ID_NO_FEATURED_IMAGE_SET' const to 'image/constants.js'