[RNMobile] Tweaks to Badge Component's Styling (#32865)
* Update background colour

The background colour of the badge component is updated to a darker gray colour with this commit (from $gray-70 to $gray-dark). This matches the colour of the edit button, which is displayed to the upper right of the image block when selected.

* Update margin and padding

The component's margin and padding is updated with this commit to match the original designs for the badge component.

* Update CHANGELOG.md

* Revert margin change

It was agreed that the original margin of 8px works better for the badge, as it is more vertically aligned with the image block's edit badge. See: https://github.com/WordPress/gutenberg/pull/32865#pullrequestreview-690043181