[RNMobile] Update Loading and Failed Screens for Unsupported Block Editor (#32395)
* Add strings for use in updated layout

These strings will be referenced in an updated layout, which seeks to provide more details to users when the UBE is loading or fails to load.

* Add dimensions for use in updated layout

These dimension will be used in an update to the "foreground_view" layout, which will display when the UBE is either loading or has failed to load.

* Add 'ube_failed' image for use in new screen

This image will be used on a new screen, which will be displayed to users when the UBE fails to load.

* Add image and text to 'foreground_view'

The 'foreground_view' screen is displayed to users when the UBE is loading. With this commit, an image and text (both a title and subtitle) are added to the screen.

The image is invisible by default, as it'll only be visible if the UBE fails to load. Future commits will make the image visible in cases where the UBE doesn't load, and also change the text used for the title and subtitle.

* Declare and retrieve new views

The new views (added to 'foreground_view' in the last commit) are declared and retrieved (via 'findViewById') in this commit.

Any updates to the old 'View' have also been updated to represent the fact that 'foreground_view' is now a 'LinearLayout'.

* Introduce function that fires when UBE fails

With this commit, a 'showTroubleshootingInstructions' is introduced.

This function will fire after 10 seconds, if Gutenberg still hasn't loaded after that time.

* Make changes to progress screen when UBE fails

If the 'showTroubleshootingInstructions' function fires, this commit includes changes that will update the progress screen's title, text, and image accordingly.

* Add reference to color used in image

The "ube.failed.xml" is currently an image of a cloud, which was copied/pasted from the WordPress Android app. See: https://github.com/wordpress-mobile/WordPress-Android/blob/develop/WordPress/src/main/res/drawable/img_illustration_cloud_off_152dp.xml

The image references a "neutral_20" color and this commit adds that color to the JS side, so that it can be called correctly.

* Update CHANGELOG

* Correct typo in CHANGELOG.md

This commit corrects the reference to an "in-progress" screen to "failed".

* Optimize conditionals around "mIsGutenbergReady"

At the moment, the code that loads the failed screen in cases where "mIsGutenbergReady" is false loads in a separate if statement to code that loads if "mIsGutenbergReady" is true. This isn't optimal.

If "mIsGutenbergReady" resolves as true, then the code that loads the failed screen doesn't need to be processed at all.

This commit optimises the code by combining the current two separate if statements into an if/else statement.

* Remove unecessary text alignment

The centered text alignment was unecessary, as the text remains centered without it.

* Update CHANGELOG.md

This commit moves the description of these changes from the '1.55.2' section of the notes to the 'unreleased' section.